+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - `feed` plugin: added `authorAvatar` property [#859].
 - RAM usage info in the debugbar.
 - Starting info in the debugbar.
+- New flag `--inspect, -i` to start a insector server for debugging.
 
 ### Changed
 - The `merge` utility function changed the types to better reflect its returned type.